Add files via upload
Password Generator Program The user is allowed to enter the website, his email, and password. The program will store data provided by the user in a CSV file. Upon the user's request, the program will retrieve a password for the provided site. If the website is not in a database, the program will provide a message that a password does not exist for the given website in the database. If the user requests, the program can create a password, using the built-in password generator function.