Add files via upload
This project combines elements of OOP and API manipulation. The User is allowed to enter information about where he/she wants to travel, starting point, time of travel, desired price, and whether the flight is direct, or it has transfer airports. The PreuzmiPodatke() object then sends an API request to https://tequila-api.kiwi.com and sends information about the flight that meets the criteria via SMS to the user. For sending SMS, the code utilizes the Python Twilio method.